interview-backend/requirements.txt
livekit>=0.19.1
livekit-agents>=0.12.11
livekit-plugins-openai>=0.10.17
python-dotenv~=1.0
textwrap3
pyairtable